move filterEvents selector to selectors/event;
